+++ crates/core/config/src/project/mod.rs
@@ -1,5 +1,6 @@
 mod dep;
 mod global_config;
+mod language_platform;
 mod local_config;
 mod task;
 mod task_options;
@@ -8,6 +9,7 @@ mod workspace;
 
 pub use dep::*;
 pub use global_config::*;
+pub use language_platform::*;
 pub use local_config::*;
 pub use task::*;
 pub use task_options::*;
